feat: break and continue expressions
We can use [ break expr; ] to escape from the loop with a value. While expression returns null, but we are able to break from it and break expression might carry an expression.

This commit adds:
- bytecode generation for break and continue expressions
- rewrites the patching logic
- temporarily changes u8 to usize in opcodes to avoid crashes. This probably will be changed in the future to make the opcode enums smaller.
- add parsing logic for optional expressions. This need some more testing, because I'm not sure if optional expressions always end up with a semicolon.

+++ src/bytecode/value.rs
@@ -5,7 +5,7 @@ use derive_more::Display;
 use enum_as_inner::EnumAsInner;
 
 pub type Number = f64;
-pub type Address = u8;
+pub type Address = usize;
 
 #[derive(Debug, Display, Clone, PartialEq, PartialOrd, EnumAsInner)]
 pub enum Value {
